move beacon data normalisation to external filter and make it the deafult one
